+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- api, api-contract, types 9.6.1
 - react-{identicon, qr}, ui-{keyring, settings}, vue-identicon 2.9.12
 - api, api-contract, types 9.5.2
 - hw-ledger, keyring, util, util-crypto 10.1.11